Update readme to include functional information
So that people viewing this repository can get a quick idea of what my configuration is doing. Eventually, I would light to set this list as a series of links to various sections within the configuration so that visitors can quickly see how I am accomplishing these tasks. Since this repository is listed within the Home Assistant cookbook, I'd like to make everything as easily accessible as possible.